backend/requirements.txt
Flask==2.2.2
Flask-API==3.0.post1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Jinja2==3.1.2
mysql-connector-python==8.0.29
# mysqlclient==2.1.1
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
regex==2022.9.13
requests==2.28.1
SQLAlchemy==1.4.41
